Add version number and numerical aperture
Model input angle of photons as a perturbation of the original input angle based on the input numerical aperture. Also add a new function for generating tissue profile, used to generate new 200 x 200 x 200 tissue (fiber.bin). The top layer (z = 0) is the surface, while all the rest of the tissue is homogenous.