Add source radius and correctly implement NA angle
The new code introduces format v2, with a source radius  (plane, not sphere) and modeling of the angle of entry based on the numerical aperture. The fiber input configuration file reflects these changes. Source radius and input angle both dependent on hardcode parameters, assuming an entry vector of [0, 0, 1], but comments indicate how this can be generalized by finding orthogonal vectors.